First cut on moving example apps to dedicated js apps.
codev: isaac and lee
