+++ client-app/src/examples/news/NewsPanelItem.scss
