+++ client-app/src/examples/news/NewsPanelModel.js
